requirements.txt
matplotlib==3.10.0
python_pptx==1.0.2
seaborn==0.13.2
shap==0.46.0
streamlit==1.41.1
xgboost==2.1.3
